Simplified "parent-child" interface
Previous commits used a `React.MouseEvent` in `CounterProps` which is
shared and transmitted between the `App` component and the `Counter`
component.

This commit reduces the information to be passed from parent to child
from an entire event to only the shift state of the pointer when the
button was clicked.

This change is part of moving the responsibility for incrementing the
count from the `Counter` component to the `App` component.

Because of this change, improved the unit test to verify that, in the
case of a regular click, the shift key was *not* pressed.